+++ RazorSvelte/Pages/Index.entry.ts
@@ -2,8 +2,5 @@
 import App from "../App/index.svelte";
 
 export default new App({
-    target: document.body,
-    props: {
-        name: ".NET6, ASP.NET, Razor, Svelte, Material UI, TypeScript and SCSS"
-    }
+    target: document.body
 });